refactor issues validations and udpate imports from issues utils in shared
