+++ src/applications/appeals/10182/content/issueSummary.js
@@ -3,8 +3,9 @@ import PropTypes from 'prop-types';
 import { Link } from 'react-router';
 
 import { ShowIssuesList } from '../components/ShowIssuesList';
-import { getSelected } from '../../shared/utils/issues';
+
 import { CONTESTABLE_ISSUES_PATH } from '../../shared/constants';
+import { getSelected } from '../../shared/utils/issues';
 
 export const SummaryTitle = ({ formData }) => {
   const issues = getSelected(formData);
