+++ src/applications/appeals/995/components/EvidenceVaRecords.jsx
@@ -12,11 +12,9 @@ import {
 import debounce from 'platform/utilities/data/debounce';
 
 import { EVIDENCE_VA_PATH, NO_ISSUES_SELECTED } from '../constants';
-
 import { content } from '../content/evidenceVaRecords';
-import { getSelected, getIssueName } from '../utils/helpers';
+import { getIssueName } from '../utils/helpers';
 import { getIndex, hasErrors } from '../utils/evidence';
-
 import { checkValidations } from '../validations';
 import {
   validateVaLocation,
@@ -27,6 +25,8 @@ import {
   isEmptyVaEntry,
 } from '../validations/evidence';
 import { focusEvidence } from '../utils/focus';
+
+import { getSelected } from '../../shared/utils/issues';
 
 const VA_PATH = `/${EVIDENCE_VA_PATH}`;
 // const REVIEW_AND_SUBMIT = '/review-and-submit';
