+++ src/applications/appeals/995/containers/ConfirmationPage.jsx
@@ -8,10 +8,11 @@ import scrollTo from 'platform/utilities/ui/scrollTo';
 import { waitForRenderThenFocus } from 'platform/utilities/ui';
 import { resetStoredSubTask } from 'platform/forms/sub-task';
 
-import { getSelected, getIssueName } from '../utils/helpers';
+import { getIssueName } from '../utils/helpers';
 import GetFormHelp from '../content/GetFormHelp';
 
 import { FORMAT_READABLE } from '../../shared/constants';
+import { getSelected } from '../../shared/utils/issues';
 
 export const ConfirmationPage = () => {
   const alertRef = useRef(null);
