+++ src/applications/appeals/996/components/ContestableIssuesWidget.jsx
@@ -2,8 +2,8 @@ import React, { useEffect, useState } from 'react';
 import PropTypes from 'prop-types';
 import { connect } from 'react-redux';
 import { Link } from 'react-router';
-import { VaModal } from '@department-of-veterans-affairs/component-library/dist/react-bindings';
 
+import { VaModal } from '@department-of-veterans-affairs/component-library/dist/react-bindings';
 import set from 'platform/utilities/data/set';
 import { setData } from 'platform/forms-system/src/js/actions';
 
@@ -18,18 +18,18 @@ import {
   removeModalContent,
 } from '../../shared/content/contestableIssues';
 import {
-  getSelected,
-  someSelected,
-  calculateIndexOffset,
-} from '../../shared/utils/issues';
-import { isEmptyObject } from '../../shared/utils/helpers';
-import { focusIssue } from '../../shared/utils/focus';
-import {
   LAST_ISSUE,
   MAX_LENGTH,
   REVIEW_ISSUES,
   SELECTED,
 } from '../../shared/constants';
+import { focusIssue } from '../../shared/utils/focus';
+import { isEmptyObject } from '../../shared/utils/helpers';
+import {
+  getSelected,
+  someSelected,
+  calculateIndexOffset,
+} from '../../shared/utils/issues';
 
 /**
  * ContestableIssuesWidget - Form system parameters passed into this widget
