+++ src/applications/appeals/996/containers/ConfirmationPage.jsx
@@ -9,9 +9,10 @@ import scrollToTop from 'platform/utilities/ui/scrollToTop';
 import { CONTACTS } from '@department-of-veterans-affairs/component-library/contacts';
 
 import { SAVED_CLAIM_TYPE, WIZARD_STATUS } from '../constants';
-import { getSelected, getIssueName } from '../utils/helpers';
+import { getIssueName } from '../utils/helpers';
 
 import { FORMAT_READABLE } from '../../shared/constants';
+import { getSelected } from '../../shared/utils/issues';
 
 export class ConfirmationPage extends React.Component {
   componentDidMount() {
