+++ src/applications/appeals/996/content/issueSummary.jsx
@@ -2,9 +2,10 @@ import React from 'react';
 import { Link } from 'react-router';
 import PropTypes from 'prop-types';
 
-import { getSelected } from '../utils/helpers';
 import { ShowIssuesList } from '../components/ShowIssuesList';
+
 import { CONTESTABLE_ISSUES_PATH } from '../../shared/constants';
+import { getSelected } from '../../shared/utils/issues';
 
 export const SummaryTitle = ({ formData }) => {
   const issues = getSelected(formData);
