+++ src/applications/appeals/996/pages/contestableIssues.js
@@ -1,10 +1,10 @@
 import ContestableIssuesWidget from '../components/ContestableIssuesWidget';
-
 import { disabilitiesExplanation } from '../content/contestableIssues';
+import { maxIssues } from '../validations/issues';
 
-import { selectionRequired, maxIssues } from '../validations/issues';
-import { hasSomeSelected } from '../utils/helpers';
 import { SELECTED } from '../../shared/constants';
+import { hasSomeSelected } from '../../shared/utils/issues';
+import { selectionRequired } from '../../shared/validations/issue';
 
 /**
  * contestable issues with add issue link (list loop)
