+++ src/applications/appeals/996/tests/validations/issues.unit.spec.js
@@ -6,12 +6,12 @@ import {
   uniqueIssue,
   maxIssues,
   areaOfDisagreementRequired,
-  selectionRequired,
   missingIssueName,
   maxNameLength,
 } from '../../validations/issues';
 
 import { MAX_LENGTH, SELECTED } from '../../../shared/constants';
+import { selectionRequired } from '../../../shared/validations/issue';
 
 describe('uniqueIssue', () => {
   const _ = null;
