+++ src/applications/appeals/996/utils/helpers.js
@@ -1,8 +1,11 @@
 import moment from 'moment';
 
 import { LEGACY_TYPE } from '../constants';
-import { processContestableIssues } from '../../shared/utils/issues';
-import { SELECTED } from '../../shared/constants';
+
+import {
+  getSelected,
+  processContestableIssues,
+} from '../../shared/utils/issues';
 import '../../shared/definitions';
 
 /**
@@ -69,26 +72,6 @@ export const mayHaveLegacyAppeals = ({
   legacyCount = 0,
   additionalIssues,
 } = {}) => legacyCount > 0 || additionalIssues?.length > 0;
-
-export const someSelected = issues =>
-  (issues || []).some(issue => issue[SELECTED]);
-
-export const hasSomeSelected = ({ contestedIssues, additionalIssues } = {}) =>
-  someSelected(contestedIssues) || someSelected(additionalIssues);
-
-export const getSelected = formData => {
-  const eligibleIssues = (formData?.contestedIssues || []).filter(
-    issue => issue[SELECTED],
-  );
-  const addedIssues = (formData?.additionalIssues || []).filter(
-    issue => issue[SELECTED],
-  );
-  // include index to help with error messaging
-  return [...eligibleIssues, ...addedIssues].map((issue, index) => ({
-    ...issue,
-    index,
-  }));
-};
 
 // additionalIssues (items) are separate because we're checking the count before
 // the formData is updated
